reorganize repo so executables are separate from libraries
